feat: page restructure
