+++ web/pages/drops/[id]/pdf.tsx
